backend/requirements.txt
asgiref==3.11.1
Django==5.1.5
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.1.0
psycopg2==2.9.11
sqlparse==0.5.5
tzdata==2025.3
